♻️ Move environment variables to new module
